Create sr-context lib
