+++ libs/sr-context/README.md
@@ -1,0 +1,7 @@
+# sr-context
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test sr-context` to execute the unit tests via [Jest](https://jestjs.io).
