+++ libs/sr-context/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "sr-context",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/sr-context/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/sr-context/**/*.{ts,tsx,js,jsx}"]
+      }
+    }
+  }
+}
